+++ src/Makefile.am
@@ -138,6 +138,7 @@ BITCOIN_CORE_H = \
   utilmoneystr.h \
   utilstrencodings.h \
   utiltime.h \
+	validationinterface.h \
   version.h \
   wallet/walletdb.h \
   wwallet/wallet.h \
@@ -188,6 +189,7 @@ libbitcoin_server_a_SOURCES = \
   timedata.cpp \
   txdb.cpp \
   txmempool.cpp \
+	validationinterface.cpp \
   $(JSON_H) \
   $(BITCOIN_CORE_H)
