+++ src/wallet/wallet.h
@@ -7,13 +7,16 @@
 #define BITCOIN_WALLET_H
 
 #include "amount.h"
-#include "primitives/block.h"
-#include "primitives/transaction.h"
 #include "tinyformat.h"
 #include "crypter.h"
 #include "key.h"
 #include "keystore.h"
+#include "primitives/block.h"
+#include "primitives/transaction.h"
+#include "tinyformat.h"
 #include "ui_interface.h"
+#include "utilstrencodings.h"
+#include "validationinterface.h"
 #include "utilstrencodings.h"
 #include "wallet/wallet_ismine.h"
 #include "wallet/walletdb.h"
